pyproject.toml
[tool.poetry]
name = "repleet-backend"
version = "0.1.0"
description = ""
authors = ["Ezekiel Victor <codeepoch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
django = "5.1.1"
djangorestframework = "3.15.2"
sqlparse = "0.5.1"
tzdata = "2024.2"
numpy = "<2.0"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
orjson = "^3.10.7"
celery = "^5.4.0"
redis = "^5.1.1"
google-generativeai = "^0.8.3"
scipy = "^1.14.1"
openai = "^1.53.0"
scikit-learn = "^1.5.2"
python-decouple = "^3.8"
dtaidistance = "^2.3.12"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
django-cors-headers = "^4.6.0"
pip-tools = "^7.4.1"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

